refactor: change endpoint from POST to GET
